Add files via upload
Save the HTML code in a file named index.html, the CSS code in a file named style.css, and the JavaScript code in a file named script.js. Place all these files in the same directory, and open the HTML file in a web browser. You should see a temperature converter where you can enter a temperature and select the input unit (Celsius or Fahrenheit). After clicking the "Convert" button, the converted temperature will be displayed below.

Signed-off-by: SaurabhKumarSingh8818 <111280762+SaurabhKumarSingh8818@users.noreply.github.com>